changed to ensure zone is a float so it always has a decimal

+++ GIS/grids/make_grids.py
@@ -87,14 +87,14 @@ def getFeatures(gdf):
 
 ZONE_MIN = 15 + (settings.longitude_min + 93.0) / 6.0
 if int(ZONE_MIN) + 0.5 > ZONE_MIN:
-    ZONE_MIN = int(ZONE_MIN)
+    ZONE_MIN = float(int(ZONE_MIN))
 else:
     ZONE_MIN = int(ZONE_MIN) + 0.5
 ZONE_MAX = 15 + (settings.longitude_max + 93.0) / 6.0
 if round(ZONE_MAX, 0) < ZONE_MAX:
     ZONE_MAX = int(ZONE_MAX) + 0.5
 else:
-    ZONE_MAX = round(ZONE_MAX, 0)
+    ZONE_MAX = float(round(ZONE_MAX, 0))
 if not os.path.exists(DIR):
     os.makedirs(DIR)
 if not os.path.exists(TMP):
